Implement formal "run Bazel cmd, get res as string"
With this function, we replace a bunch of hacking left and right involving
injecting the `shell-command-to-string function along with requisite Bazel
flags, e.g. "--noshow_loading_progress", to get the bare-bones string result
that we expect. Just leaky abstractions all around. What's more, now we have a
central place for all of such requisite flags, along with the abysmal `| cat`
hack...